Lib proxi mng (#196)
* Change Front and back proxi init

* Fix linter errors

* Change calibration from drivers to manager for imus

* Change input files and made Fake IMU only one instance

* Fix warnings

* Fix warnings

+++ BeagleBone_black/src/sensors/fake_gpio_counter.cpp
@@ -63,7 +63,7 @@ void FakeGpioCounter::readDataFromFile(std::string file_path)
     }
 
     uint32_t temp_value;
-    int counter = 0;
+    uint64_t counter = 0;
     uint32_t temp_time;
     std::string line;
